Moving files into what I think will be a better structure for the future
