+++ app/src/app/components/explorer/path-change/PathChangeContainer.ts
@@ -1,9 +1,9 @@
-import { INotepadsStoreState, INotepadStoreState } from '../types/NotepadTypes';
-import { IStoreState } from '../types';
+import { INotepadsStoreState, INotepadStoreState } from '../../../types/NotepadTypes';
+import { IStoreState } from '../../../types';
 import { connect } from 'react-redux';
-import PathChangeComponent, { IPathChangeComponentProps } from '../components/explorer/PathChangeComponent';
+import PathChangeComponent, { IPathChangeComponentProps } from './PathChangeComponent';
 import { Action, Dispatch } from 'redux';
-import { actions } from '../actions';
+import { actions } from '../../../actions';
 
 export function mapStateToProps({ notepads }: IStoreState): IPathChangeComponentProps {
 	const notepad = ((notepads || {} as INotepadsStoreState).notepad || {} as INotepadStoreState).item;
